requirements.txt
# Core dependencies
pywin32==306
mss==9.0.1
opencv-python==4.8.1.78
pytesseract==0.3.10
pyautogui==0.9.54
pynput==1.7.6
Pillow==10.1.0
numpy==1.26.2  # Updated for Python 3.12 compatibility

# AI and Web
anthropic==0.18.0
requests==2.31.0
beautifulsoup4==4.12.2

# Performance and monitoring
psutil==5.9.6
cachetools==5.3.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Building
pyinstaller==6.3.0

# UI enhancements
customtkinter==5.2.1
pystray==0.19.5
plyer==2.1
matplotlib==3.8.2  # Updated for Python 3.12 compatibility

# Machine Learning (Optional - comment out if not needed)
scikit-learn==1.3.2
scipy==1.11.4  # Updated for Python 3.12 compatibility
# torch==2.1.0  # Comment out - requires separate installation
# torchvision==0.16.0  # Comment out - requires separate installation
packaging==23.2

# Localization
babel==2.13.1
pytz==2023.3

# Security (for encryption features)
cryptography==41.0.5
keyring==24.3.0